ai_data_analyst/src/requirements.txt
streamlit>=1.37.0
pandas>=2.0.0
plotly>=5.20.0
openai>=1.0.0
python-dotenv>=1.0.1
numpy>=1.24.0
